Refactor into one file
