+++ packages/bruno-app/src/components/Sidebar/Collections/Collection/CollectionItem/GenerateCodeItem/index.js
@@ -5,8 +5,7 @@ import StyledWrapper from './StyledWrapper';
 import { isValidUrl } from 'utils/url/index';
 import get from 'lodash/get';
 import { findEnvironmentInCollection } from 'utils/collections';
-const { interpolateVars, interpolateUrl } = require('@usebruno/js/src/interpolate-vars');
-const { interpolateString } = require('@usebruno/js/src/interpolate-string');
+const { interpolateVars, interpolateUrl, interpolateString } = require('@usebruno/js/src/interpolate');
 
 const languages = [
   {
